@@ -1,9 +1,8 @@
 <?php
+declare(strict_types=1);
 
 use App\Kernel;
 
-require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
+require_once dirname(__DIR__) . '/vendor/autoload_runtime.php';
 
-return function (array $context) {
-    return new Kernel($context['APP_ENV'], (bool) $context['APP_DEBUG']);
-};
+return static fn (array $context) => new Kernel($context['APP_ENV'], (bool) $context['APP_DEBUG']);
